finn/requirements.txt
openpyxl==3.0.10
pandas==1.4.4
Pillow==9.4.0
streamlit==1.16.0
streamlit_authenticator==0.2.1
